refactor: profile pagae ssr & debug: recruitList
