+++ src/app/(provider)/_providers/auth.provider.ts
@@ -14,7 +14,6 @@ function AuthProvider({ children }: PropsWithChildren) {
 
   useEffect(() => {
     supabase.auth.onAuthStateChange((_name, session) => {
-      console.log(session);
       if (session) {
         const userId = session.user.id;
         setCurrentUserId(userId);
